requirements.txt
fastapi
murf
google-genai
assemblyai
dotenv
uvicorn
python-multipart
websocket
websockets
logging
tavily-python
attrs
beautifulsoup4
pydantic